refactor: move fee calculation to `fee` field type
