+++ src/tx/builder/constants.ts
@@ -95,3 +95,58 @@ export const ID_TAG_PREFIX = mapObject(
   PREFIX_ID_TAG,
   ([key, value]: [EncodingType, number]) => [value, key],
 );
+
+/**
+ * Enum with transaction types
+ * @category transaction builder
+ * @see {@link https://github.com/aeternity/protocol/blob/0f6dee3d9d1e8e2469816798f5c7587a6c918f94/serializations.md#binary-serialization}
+ */
+export enum TX_TYPE {
+  account = 10,
+  signed = 11,
+  spend = 12,
+  oracleRegister = 22,
+  oracleQuery = 23,
+  oracleResponse = 24,
+  oracleExtend = 25,
+  nameClaim = 32,
+  namePreClaim = 33,
+  nameUpdate = 34,
+  nameRevoke = 35,
+  nameTransfer = 36,
+  contract = 40,
+  contractCallResult = 41,
+  contractCreate = 42,
+  contractCall = 43,
+  channelCreate = 50,
+  channelDeposit = 51,
+  channelWithdraw = 52,
+  channelCloseMutual = 53,
+  channelCloseSolo = 54,
+  channelSlash = 55,
+  channelSettle = 56,
+  channelOffChain = 57,
+  channel = 58,
+  channelSnapshotSolo = 59,
+  proofOfInclusion = 60,
+  stateTrees = 62,
+  merklePatriciaTree = 63,
+  merklePatriciaTreeValue = 64,
+  sophiaByteCode = 70,
+  gaAttach = 80,
+  gaMeta = 81,
+  payingFor = 82,
+  channelForceProgress = 521,
+  channelOffChainUpdateTransfer = 570,
+  channelOffChainUpdateDeposit = 571,
+  channelOffChainUpdateWithdrawal = 572,
+  channelOffChainCreateContract = 573,
+  channelOffChainCallContract = 574,
+  channelReconnect = 575,
+  contractsTree = 621,
+  contractCallsTree = 622,
+  channelsTree = 623,
+  nameserviceTree = 624,
+  oraclesTree = 625,
+  accountsTree = 626,
+}
