+++ src/tx/builder/schema.ts
@@ -5,8 +5,9 @@
 // # https://github.com/aeternity/protocol/blob/master/serializations.md#binary-serialization
 
 import BigNumber from 'bignumber.js';
+import { TX_TYPE } from './constants';
 import {
-  Field, uInt, shortUInt, coinAmount, name, nameId, nameFee, deposit, gasPrice,
+  Field, uInt, shortUInt, coinAmount, name, nameId, nameFee, deposit, gasPrice, fee,
 } from './field-types';
 import { EncodedData, EncodingType } from '../../utils/encoder';
 import MPTree from '../../utils/mptree';
@@ -40,61 +41,6 @@ export type TxField = [
   type: FIELD_TYPES | Field,
   prefix?: EncodingType | EncodingType[],
 ];
-
-/**
- * Enum with transaction types
- * @category transaction builder
- * @see {@link https://github.com/aeternity/protocol/blob/0f6dee3d9d1e8e2469816798f5c7587a6c918f94/serializations.md#binary-serialization}
- */
-export enum TX_TYPE {
-  account = 10,
-  signed = 11,
-  spend = 12,
-  oracleRegister = 22,
-  oracleQuery = 23,
-  oracleResponse = 24,
-  oracleExtend = 25,
-  nameClaim = 32,
-  namePreClaim = 33,
-  nameUpdate = 34,
-  nameRevoke = 35,
-  nameTransfer = 36,
-  contract = 40,
-  contractCallResult = 41,
-  contractCreate = 42,
-  contractCall = 43,
-  channelCreate = 50,
-  channelDeposit = 51,
-  channelWithdraw = 52,
-  channelCloseMutual = 53,
-  channelCloseSolo = 54,
-  channelSlash = 55,
-  channelSettle = 56,
-  channelOffChain = 57,
-  channel = 58,
-  channelSnapshotSolo = 59,
-  proofOfInclusion = 60,
-  stateTrees = 62,
-  merklePatriciaTree = 63,
-  merklePatriciaTreeValue = 64,
-  sophiaByteCode = 70,
-  gaAttach = 80,
-  gaMeta = 81,
-  payingFor = 82,
-  channelForceProgress = 521,
-  channelOffChainUpdateTransfer = 570,
-  channelOffChainUpdateDeposit = 571,
-  channelOffChainUpdateWithdrawal = 572,
-  channelOffChainCreateContract = 573,
-  channelOffChainCallContract = 574,
-  channelReconnect = 575,
-  contractsTree = 621,
-  contractCallsTree = 622,
-  channelsTree = 623,
-  nameserviceTree = 624,
-  oraclesTree = 625,
-  accountsTree = 626,
-}
 
 /**
  * @category transaction builder
@@ -243,7 +189,7 @@ type BuildTxArgBySchema<SchemaLine> =
   >;
 
 export type RawTxObject<Tx extends TxSchema> = {
-  [k in keyof Tx]: Tx[k] extends BigNumber ? string : Tx[k]
+  [k in keyof Tx]-?: Tx[k] extends BigNumber ? string : Tx[k]
 };
 
 const BASE_TX = [
@@ -275,7 +221,7 @@ export const TX_SCHEMA = {
       ['senderId', FIELD_TYPES.id, 'ak'],
       ['recipientId', FIELD_TYPES.id, ['ak', 'nm']],
       ['amount', coinAmount],
-      ['fee', uInt],
+      ['fee', fee],
       ['ttl', shortUInt],
       ['nonce', shortUInt],
       ['payload', FIELD_TYPES.payload],
@@ -287,7 +233,7 @@ export const TX_SCHEMA = {
       ['accountId', FIELD_TYPES.id, 'ak'],
       ['nonce', shortUInt],
       ['commitmentId', FIELD_TYPES.id, 'cm'],
-      ['fee', uInt],
+      ['fee', fee],
       ['ttl', shortUInt],
     ],
   },
@@ -299,7 +245,7 @@ export const TX_SCHEMA = {
       ['name', name],
       ['nameSalt', uInt],
       ['nameFee', nameFee],
-      ['fee', uInt],
+      ['fee', fee],
       ['ttl', shortUInt],
     ],
   },
@@ -312,7 +258,7 @@ export const TX_SCHEMA = {
       ['nameTtl', uInt],
       ['pointers', FIELD_TYPES.pointers],
       ['clientTtl', shortUInt],
-      ['fee', uInt],
+      ['fee', fee],
       ['ttl', shortUInt],
     ],
   },
@@ -323,7 +269,7 @@ export const TX_SCHEMA = {
       ['nonce', shortUInt],
       ['nameId', nameId],
       ['recipientId', FIELD_TYPES.id, ['ak', 'nm']],
-      ['fee', uInt],
+      ['fee', fee],
       ['ttl', shortUInt],
     ],
   },
@@ -333,7 +279,7 @@ export const TX_SCHEMA = {
       ['accountId', FIELD_TYPES.id, 'ak'],
       ['nonce', shortUInt],
       ['nameId', nameId],
-      ['fee', uInt],
+      ['fee', fee],
       ['ttl', shortUInt],
     ],
   },
@@ -356,7 +302,7 @@ export const TX_SCHEMA = {
       ['nonce', shortUInt],
       ['code', FIELD_TYPES.binary, 'cb'],
       ['ctVersion', FIELD_TYPES.ctVersion],
-      ['fee', uInt],
+      ['fee', fee],
       ['ttl', shortUInt],
       ['deposit', deposit],
       ['amount', coinAmount],
@@ -372,7 +318,7 @@ export const TX_SCHEMA = {
       ['nonce', shortUInt],
       ['contractId', FIELD_TYPES.id, ['ct', 'nm']],
       ['abiVersion', FIELD_TYPES.abiVersion],
-      ['fee', uInt],
+      ['fee', fee],
       ['ttl', shortUInt],
       ['amount', coinAmount],
       ['gasLimit', shortUInt],
@@ -406,7 +352,7 @@ export const TX_SCHEMA = {
       ['queryFee', coinAmount],
       ['oracleTtlType', FIELD_TYPES.ttlType],
       ['oracleTtlValue', shortUInt],
-      ['fee', uInt],
+      ['fee', fee],
       ['ttl', shortUInt],
       ['abiVersion', FIELD_TYPES.abiVersion],
     ],
@@ -418,7 +364,7 @@ export const TX_SCHEMA = {
       ['nonce', shortUInt],
       ['oracleTtlType', FIELD_TYPES.ttlType],
       ['oracleTtlValue', shortUInt],
-      ['fee', uInt],
+      ['fee', fee],
       ['ttl', shortUInt],
     ],
   },
@@ -434,7 +380,7 @@ export const TX_SCHEMA = {
       ['queryTtlValue', shortUInt],
       ['responseTtlType', FIELD_TYPES.ttlType],
       ['responseTtlValue', shortUInt],
-      ['fee', uInt],
+      ['fee', fee],
       ['ttl', shortUInt],
     ],
   },
@@ -447,7 +393,7 @@ export const TX_SCHEMA = {
       ['response', FIELD_TYPES.string],
       ['responseTtlType', FIELD_TYPES.ttlType],
       ['responseTtlValue', shortUInt],
-      ['fee', uInt],
+      ['fee', fee],
       ['ttl', shortUInt],
     ],
   },
@@ -461,7 +407,7 @@ export const TX_SCHEMA = {
       ['channelReserve', uInt],
       ['lockPeriod', uInt],
       ['ttl', shortUInt],
-      ['fee', uInt],
+      ['fee', fee],
       ['initiatorDelegateIds', FIELD_TYPES.string],
       ['responderDelegateIds', FIELD_TYPES.string],
       ['stateHash', FIELD_TYPES.binary, 'st'],
@@ -476,7 +422,7 @@ export const TX_SCHEMA = {
       ['initiatorAmountFinal', uInt],
       ['responderAmountFinal', uInt],
       ['ttl', shortUInt],
-      ['fee', uInt],
+      ['fee', fee],
       ['nonce', shortUInt],
     ],
   },
@@ -488,7 +434,7 @@ export const TX_SCHEMA = {
       ['payload', FIELD_TYPES.binary, 'tx'],
       ['poi', FIELD_TYPES.binary, 'pi'],
       ['ttl', shortUInt],
-      ['fee', uInt],
+      ['fee', fee],
       ['nonce', shortUInt],
     ],
   },
@@ -500,7 +446,7 @@ export const TX_SCHEMA = {
       ['payload', FIELD_TYPES.binary, 'tx'],
       ['poi', FIELD_TYPES.binary, 'pi'],
       ['ttl', shortUInt],
-      ['fee', uInt],
+      ['fee', fee],
       ['nonce', shortUInt],
     ],
   },
@@ -511,7 +457,7 @@ export const TX_SCHEMA = {
       ['fromId', FIELD_TYPES.id, 'ak'],
       ['amount', uInt],
       ['ttl', shortUInt],
-      ['fee', uInt],
+      ['fee', fee],
       ['stateHash', FIELD_TYPES.binary, 'st'],
       ['round', uInt],
       ['nonce', shortUInt],
@@ -524,7 +470,7 @@ export const TX_SCHEMA = {
       ['toId', FIELD_TYPES.id, 'ak'],
       ['amount', uInt],
       ['ttl', shortUInt],
-      ['fee', uInt],
+      ['fee', fee],
       ['stateHash', FIELD_TYPES.binary, 'st'],
       ['round', uInt],
       ['nonce', shortUInt],
@@ -538,7 +484,7 @@ export const TX_SCHEMA = {
       ['initiatorAmountFinal', uInt],
       ['responderAmountFinal', uInt],
       ['ttl', shortUInt],
-      ['fee', uInt],
+      ['fee', fee],
       ['nonce', shortUInt],
     ],
   },
@@ -553,7 +499,7 @@ export const TX_SCHEMA = {
       ['stateHash', FIELD_TYPES.binary, 'st'],
       ['offChainTrees', FIELD_TYPES.stateTree],
       ['ttl', shortUInt],
-      ['fee', uInt],
+      ['fee', fee],
       ['nonce', shortUInt],
     ],
   },
@@ -592,7 +538,7 @@ export const TX_SCHEMA = {
       ['fromId', FIELD_TYPES.id, 'ak'],
       ['payload', FIELD_TYPES.binary, 'tx'],
       ['ttl', shortUInt],
-      ['fee', uInt],
+      ['fee', fee],
       ['nonce', shortUInt],
     ],
   },
@@ -729,7 +675,7 @@ export const TX_SCHEMA = {
       ['code', FIELD_TYPES.binary, 'cb'],
       ['authFun', FIELD_TYPES.rawBinary],
       ['ctVersion', FIELD_TYPES.ctVersion],
-      ['fee', uInt],
+      ['fee', fee],
       ['ttl', shortUInt],
       ['gasLimit', shortUInt],
       ['gasPrice', gasPrice],
@@ -742,7 +688,7 @@ export const TX_SCHEMA = {
       ['gaId', FIELD_TYPES.id, 'ak'],
       ['authData', FIELD_TYPES.binary, 'cb'],
       ['abiVersion', FIELD_TYPES.abiVersion],
-      ['fee', uInt],
+      ['fee', fee],
       ['gasLimit', shortUInt],
       ['gasPrice', gasPrice],
       ['tx', FIELD_TYPES.rlpBinary],
@@ -753,7 +699,7 @@ export const TX_SCHEMA = {
       ...BASE_TX,
       ['payerId', FIELD_TYPES.id, 'ak'],
       ['nonce', shortUInt],
-      ['fee', uInt],
+      ['fee', fee],
       ['tx', FIELD_TYPES.rlpBinary],
     ],
   },
