+++ src/tx/validator.ts
@@ -5,7 +5,6 @@ import {
   PROTOCOL_VM_ABI, RawTxObject, TxSchema, TxParamsCommon, TX_TYPE, TxTypeSchemas, CtVersion,
 } from './builder/schema';
 import { TxUnpacked, unpackTx } from './builder';
-import calculateMinFee from './min-fee';
 import { UnsupportedProtocolError } from '../utils/errors';
 import { concatBuffers, isKeyOfObject } from '../utils/other';
 import { EncodedData } from '../utils/encoder';
@@ -130,16 +129,6 @@ validators.push(
       node,
       [...parentTxTypes, txType],
     );
-  },
-  (tx: any, { txType }) => {
-    if (tx.fee === undefined) return [];
-    const minFee = calculateMinFee(txType, { params: tx, vsn: tx.VSN });
-    if (minFee.lte(tx.fee)) return [];
-    return [{
-      message: `Fee ${tx.fee as string} is too low, minimum fee for this transaction is ${minFee.toString()}`,
-      key: 'InsufficientFee',
-      checkedKeys: ['fee'],
-    }];
   },
   ({ ttl }, { height }) => {
     if (ttl === undefined) return [];
